pull out custom character type to own file
