+++ src/customTypes.tsx
@@ -1,0 +1,19 @@
+type Character = {
+  id: number;
+  name: string;
+  dob: string;
+  personality: string;
+  appearance: string;
+  background: string;
+  gender: string;
+  race: string;
+  ethnicity: string;
+  source: string;
+  profile: string;
+  relationships: string[];
+  images: string[];
+  dislikes: string[];
+  likes: string[];
+};
+
+export default Character;
