+++ src/routes/ViewAll.tsx
@@ -3,25 +3,8 @@ import { Container } from "react-bootstrap";
 import ReactLoading from "react-loading";
 import Row from "react-bootstrap/Row";
 import CharacterBrief from "../components/CharacterBrief";
+import Character from "../customTypes";
 import config from "../config.json";
-
-type Character = {
-  id: number;
-  name: string;
-  dob: string;
-  personality: string;
-  appearance: string;
-  background: string;
-  gender: string;
-  race: string;
-  ethnicity: string;
-  source: string;
-  profile: string;
-  relationships: string[];
-  images: string[];
-  dislikes: string[];
-  likes: string[];
-};
 
 function ViewAll() {
   const [characters, setCharacters] = useState([]);
