+++ src/routes/ViewOne.tsx
@@ -1,25 +1,8 @@
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import ReactLoading from "react-loading";
+import Character from "../customTypes";
 import config from "../config.json";
-
-type Character = {
-  id: number;
-  name: string;
-  dob: string;
-  personality: string;
-  appearance: string;
-  background: string;
-  gender: string;
-  race: string;
-  ethnicity: string;
-  source: string;
-  profile: string;
-  relationships: string[];
-  images: string[];
-  dislikes: string[];
-  likes: string[];
-};
 
 function ViewOne() {
   const { id } = useParams();
